Setup for launching locally-hosted games with a nicer interface than writing script.txt directly.
[Incomplete] Added GameData package which contains all necessary information for setting up a game.
Added `launchScriptHostedgame(from:wantsReplayRecorded:)` to convert between the two formats.

This layer is somewhat unnecessary but will give a nicer interface to control launching games. Ideally in the future this would all use a JSON format that the engine could read directly, and the conversion would go away and simply use the new format. This would have to be sanctioned by the engine & other lobby developers, and is somewhat unlikely to occur.

Housekeeping:
- Converted `LaunchScriptWriter.filePath` and `LaunchScriptWriter.fileName` to static properties.
- Renamed the LaunchScriptWriter instance method `prepareForLaunchOfSinglePlayerGame(_:)` to `prepareForLaunchOfSpringAsHost(_:)` since the logic is the same and therefore should be shared.
- Renamed `LaunchScriptWriter.Game.startPosType` to `LaunchScriptWriter.Game.startPositionType` for readability.
- Removed the `gamename` argument from `LaunchScriptWriter`'s logic as its purpose is unclear.
- Added `Color` typealias for `UInt32` to allow easy retrieval of the red, green, and blue components.