Add temporary debug-only logger
In attempting to trace a battleroom-related bug (where status changes from ready -> unready without user request) a logger has been added to log operations when status changes and when specifically the user requests a ready change. What is logged relevant to this should be refined.
Other things can and will be logged in the future.
It is possible that this logger can be made to write logs on released builds also.