Dialog Sheets, Account Info, & MessageIDs
Added `DialogSheet` – a template class for dialog sheets that aim to gather user input for the purpose of initiating an asynchronous action (such as  changing username). `DialogSheet` uses a `.xib` file so should be copied across as a template for all instances of `DialogSheet`.

Implemented changing username, a template for changing passwords, and updated `ServerSelectionViewController` and `UserAuthenticationViewController` to be subclasses of `DialogSheet`.

Related to this – selecting server and logging in now have cancel buttons; cancelling server selection closes the window and terminates the connection, and cancelling user authentication returs to server selection.

Also added `AccountInfoViewController` (and its controller, `AccountController`) for viewing information about accounts. This relies on somewhat-questionable parsing of the `SERVERMSG` response to a `GETUSERINFO` command.
Added `MacOSConnectionWindowManager` to the responder chain (above `MainWindowController`) to allow handling of menu bar actions.

`ConnectionWindowManager` now has a reference to `ConnectionController` to allow termination of a `Connection`.

Most added commands are in `AccountMetadataCommands.swift`.

`TASServer` and `CommandHandler` now make use of messageIDs to keep track of which server command replies to which client command.

Added `RankView` subclass of `NSImageView` for transferrable conversion of a rank into an image.

Housekeeping:
- `SCServerMessage` are now processed.
- `MOTDCommand.swift` -> `GeneralServerCommands.swift`
- `ServerSelectionViewControllerDelegate -> `ServerSelectionDelegate`
- Added “Print Responders” menu item for verifying the responder chain.
- Moved `ServerAddress` to its own file.
- Added BAR Server to the list of servers.

This commit may not compile.