`Connection` renamed `Client` + Login fix
Login commands with a message ID aren’t handled correctly by Uberserver and so a fix has been provided. It can safely be assumed that the only specific handler at login time is the one listening for the response for login, and so we just call the single existing handler.

Also, Client is a better name than Connection.