Allow playback of replays
A significant amount of dependency juggling. This needs restructuring.

The lobby now can asynchronously load stored replays and show basic information about them in a view controller. `ListViewController` was used for this for simplicity, but I strongly suggest it’s not the best method. It has a wonderful loading animation though.

Lists are tagged as not thread-safe. Another convenience sorter has been provided which makes property-based sorting a little easier. This will likely replace other methods of sorting.

Packages:
- Removed decompression package, instead including the modified source directly in the codebase.

System now handles a wider range of system-specific tasks:
- Providing the window manager
- File handling
- Launching external applications

The previous implementation of start scripts was replaced in favour of a new one, and the directory was changed from the data dir to the config/cache dir.

Housekeeping:
- Removed `ListViewController`’s unused delegate & datasource APIs
- Added `NSHomeDirectoryURL() -> URL` because URLs are safer.
- Updated script password value to be more sensible

Bumped license version to GPL 3 or later for compatability with Apache.